# api/openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lisboa Pública API",
    "version": "1.0.0",
    "description": "Provider-backed Lisbon transit. apiKey scopes use x-required-scopes; all data lists are paginated. Live revisions expire with410; snapshot revisions freeze committed ingestion membership. Scheduled and estimated data are explicitly classified. Dashboard reads are public for now and rate-limited; supplied API keys still enforce x-required-scopes. Key management requires a Google session. Expensive reads share two fail-fast slots and a 15-second operation deadline (503 busy/timeout). Schedule and historical grouped lists reject over 100000 results with 400 result_limit instead of truncation; narrow operators, route or time window. Busy503 responses include Retry-After."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/v1/stops": {
      "get": {
        "operationId": "listStops",
        "summary": "listStops",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StopPage"
                }
              }
            }
          },
          "default": {
            "description": "Error (400,401,403,404,410,429,500,503). 429 and busy503 include Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "ApiKey": []
          },
          {
            "SessionCookie": []
          }
        ],
        "x-required-scopes": [
          "read:transit"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/offset"
          },
          {
            "$ref": "#/components/parameters/revision"
          },
          {
            "$ref": "#/components/parameters/operators"
          },
          {
            "$ref": "#/components/parameters/route_id"
          },
          {
            "$ref": "#/components/parameters/q"
          }
        ],
        "x-public-read": true
      }
    },
    "/api/v1/vehicles": {
      "get": {
        "operationId": "listVehicles",
        "summary": "listVehicles",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehiclePage"
                }
              }
            }
          },
          "default": {
            "description": "Error (400,401,403,404,410,429,500,503). 429 and busy503 include Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "ApiKey": []
          },
          {
            "SessionCookie": []
          }
        ],
        "x-required-scopes": [
          "read:transit"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/offset"
          },
          {
            "$ref": "#/components/parameters/revision"
          },
          {
            "$ref": "#/components/parameters/operators"
          },
          {
            "$ref": "#/components/parameters/route_id"
          },
          {
            "$ref": "#/components/parameters/stop_id"
          }
        ],
        "x-public-read": true
      }
    },
    "/api/v1/trips": {
      "get": {
        "operationId": "listTrips",
        "summary": "listTrips",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripPage"
                }
              }
            }
          },
          "default": {
            "description": "Error (400,401,403,404,410,429,500,503). 429 and busy503 include Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "ApiKey": []
          },
          {
            "SessionCookie": []
          }
        ],
        "x-required-scopes": [
          "read:transit"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/offset"
          },
          {
            "$ref": "#/components/parameters/revision"
          },
          {
            "$ref": "#/components/parameters/operators"
          },
          {
            "$ref": "#/components/parameters/route_id"
          },
          {
            "$ref": "#/components/parameters/from"
          },
          {
            "$ref": "#/components/parameters/to"
          }
        ],
        "x-public-read": true
      }
    },
    "/api/v1/arrivals": {
      "get": {
        "operationId": "listArrivals",
        "summary": "listArrivals",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArrivalPage"
                }
              }
            }
          },
          "default": {
            "description": "Error (400,401,403,404,410,429,500,503). 429 and busy503 include Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "ApiKey": []
          },
          {
            "SessionCookie": []
          }
        ],
        "x-required-scopes": [
          "read:transit"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/offset"
          },
          {
            "$ref": "#/components/parameters/revision"
          },
          {
            "$ref": "#/components/parameters/operators"
          },
          {
            "$ref": "#/components/parameters/route_id"
          },
          {
            "$ref": "#/components/parameters/from"
          },
          {
            "$ref": "#/components/parameters/to"
          },
          {
            "$ref": "#/components/parameters/stop_id"
          }
        ],
        "x-public-read": true,
        "description": "Published predictions and planned arrivals at one stop. Defaults to a one-hour window. Provider collection is shared and bounded; first reads may return loading availability with planned fallback."
      }
    },
...
    "/api/v1/history": {
      "get": {
        "operationId": "listHistory",
        "summary": "listHistory",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryPointPage"
                }
              }
            }
          },
          "default": {
            "description": "Error (400,401,403,404,410,429,500,503). 429 and busy503 include Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "ApiKey": []
          },
          {
            "SessionCookie": []
          }
        ],
        "x-required-scopes": [
          "read:history"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/offset"
          },
          {
            "$ref": "#/components/parameters/revision"
          },
          {
            "$ref": "#/components/parameters/operators"
          },
          {
            "$ref": "#/components/parameters/route_id"
          },
          {
            "$ref": "#/components/parameters/from"
          },
          {
            "$ref": "#/components/parameters/to"
          }
        ],
        "x-public-read": true
      }
    },
    "/api/v1/fleet": {
      "get": {
        "operationId": "listFleet",
        "summary": "listFleet",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FleetVehiclePage"
                }
              }
            }
          },
          "default": {
            "description": "Error (400,401,403,404,410,429,500,503). 429 and busy503 include Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "ApiKey": []
          },
          {
            "SessionCookie": []
          }
        ],
        "x-required-scopes": [
          "read:history"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/offset"
          },
          {
            "$ref": "#/components/parameters/revision"
          },
          {
            "$ref": "#/components/parameters/operators"
          },
          {
            "$ref": "#/components/parameters/route_id"
          },
          {
            "$ref": "#/components/parameters/from"
          },
          {
            "$ref": "#/components/parameters/to"
          },
          {
            "$ref": "#/components/parameters/q"
          },
          {
            "$ref": "#/components/parameters/sort"
          }
        ],
        "x-public-read": true
      }
    },
    "/api/v1/traffic": {
      "get": {
        "operationId": "listTraffic",
        "summary": "listTraffic",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrafficPointPage"
                }
              }
            }
          },
          "default": {
            "description": "Error (400,401,403,404,410,429,500,503). 429 and busy503 include Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "ApiKey": []
          },
          {
            "SessionCookie": []
          }
        ],
        "x-required-scopes": [
          "read:history"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/offset"
          },
          {
            "$ref": "#/components/parameters/revision"
          },
          {
            "$ref": "#/components/parameters/operators"
          },
          {
            "$ref": "#/components/parameters/route_id"
          },
          {
            "$ref": "#/components/parameters/from"
          },
          {
            "$ref": "#/components/parameters/to"
          },
          {
            "$ref": "#/components/parameters/hour_start"
          },
          {
            "$ref": "#/components/parameters/hour_end"
          },
          {
            "$ref": "#/components/parameters/weekdays_only"
          }
        ],
        "x-public-read": true
      }
    },
    "/api/v1/rankings": {
      "get": {
        "operationId": "listRankings",
        "summary": "listRankings",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankingPage"
                }
              }
            }
          },
          "default": {
            "description": "Error (400,401,403,404,410,429,500,503). 429 and busy503 include Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "ApiKey": []
          },
          {
            "SessionCookie": []
          }
        ],
        "x-required-scopes": [
          "read:history"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/offset"
          },
          {
            "$ref": "#/components/parameters/revision"
          },
          {
            "$ref": "#/components/parameters/operators"
          },
          {
            "$ref": "#/components/parameters/route_id"
          },
          {
            "$ref": "#/components/parameters/from"
          },
          {
            "$ref": "#/components/parameters/to"
          },
          {
            "$ref": "#/components/parameters/sort"
          }
        ],
        "x-public-read": true
      }
    },
...
    "/api/v1/operator-coverage": {
      "get": {
        "operationId": "listOperatorCoverage",
        "summary": "Committed per-operator historical coverage",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorCoveragePage"
                }
              }
            }
          },
          "default": {
            "description": "Error (400,401,403,404,410,429,500,503). 429 and busy503 include Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "ApiKey": []
          },
          {
            "SessionCookie": []
          }
        ],
        "x-required-scopes": [
          "read:history"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/offset"
          },
          {
            "$ref": "#/components/parameters/revision"
          },
          {
            "$ref": "#/components/parameters/operators"
          },
          {
            "$ref": "#/components/parameters/route_id"
          },
          {
            "$ref": "#/components/parameters/from"
          },
          {
            "$ref": "#/components/parameters/to"
          },
          {
            "$ref": "#/components/parameters/hour_start"
          },
          {
            "$ref": "#/components/parameters/hour_end"
          },
          {
            "$ref": "#/components/parameters/weekdays_only"
          }
        ],
        "x-public-read": true
      }
    },
    "/api/v1/cp/predictions": {
      "get": {
        "operationId": "listCpPredictions",
        "summary": "Public CP stop-call predictions with original clocks and partial coverage",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CPPredictionPage"
                }
              }
            }
          },
          "default": {
            "description": "Error (400,401,403,404,410,429,500,503). 429 and busy503 include Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "ApiKey": []
          },
          {
            "SessionCookie": []
          }
        ],
        "x-required-scopes": [
          "read:transit"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/offset"
          },
          {
            "$ref": "#/components/parameters/revision"
          },
          {
            "$ref": "#/components/parameters/operators"
          },
          {
            "$ref": "#/components/parameters/route_id"
          },
          {
            "$ref": "#/components/parameters/from"
          },
          {
            "$ref": "#/components/parameters/to"
          },
          {
            "$ref": "#/components/parameters/stop_id"
          },
          {
            "name": "trip_id",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 256
            },
            "description": "Source-qualified GTFS trip ID; never a generated dated arrival ID."
          }
        ],
        "x-public-read": true
      }
    },
...
    "/api/v1/metro/patterns": {
      "get": {
        "operationId": "getMetroPatterns",
        "summary": "Experimental Metro patterns and official/own forecasts; no physical validation",
        "security": [
          {},
          {
            "ApiKey": []
          },
          {
            "SessionCookie": []
          }
        ],
        "x-public-read": true,
        "x-required-scopes": [
          "read:transit",
          "read:history"
        ],
        "responses": {
          "200": {
            "description": "Partial data and unavailable metrics remain explicit; collection status includes archive failures.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetroPatterns"
                }
              }
            }
          },
          "default": {
            "description": "Error (400,401,403,404,410,429,500,503). 429 and busy503 include Retry-After.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "stop_id",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 240
            }
          },
          {
            "name": "episode",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^[a-f0-9]{24}$"
            }
          }
        ]
      }
    },
    "/api/v1/transport/patterns": {
      "get": {
        "operationId": "getTransportPatterns",
        "summary": "Experimental transport patterns and separate official/own forecasts",
        "security": [
          {},
          {
            "ApiKey": []
          },
          {
            "SessionCookie": []
          }
        ],
        "x-public-read": true,
        "x-required-scopes": [
          "read:transit",
          "read:history"
        ],
        "responses": {
          "200": {
            "description": "Partial data and unavailable metrics remain explicit; collection status includes archive failures.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetroPatterns"
                }
              }
            }
          },
          "default": {
            "description": "Error (400,401,403,404,410,429,500,503). 429 and busy503 include Retry-After.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "operator_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "metro",
                "cm",
                "carris",
                "cp",
                "fertagus",
                "ttsl",
                "tcb",
                "mobi"
              ]
            }
          },
          {
            "name": "stop_id",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 240
            }
          },
          {
            "name": "episode",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^[a-f0-9]{24}$"
            }
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Bearer lp_... Scoped read:transit or read:history; requirements are x-required-scopes."
      },
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "lp_session"
      }
    },
    "parameters": {
      "limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 500,
          "default": 100
        }
      },
      "offset": {
        "name": "offset",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 0,
          "maximum": 1000000,
          "default": 0
        }
      },
      "revision": {
        "name": "revision",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 120
        },
        "description": "Opaque collection revision from page.revision; reuse on later pages. Schedule revisions also freeze their time window and prediction eligibility. Expired revisions return410."
      },
      "operators": {
        "name": "operators",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 200,
          "description": "Comma-separated operator IDs."
        }
      },
      "route_id": {
        "name": "route_id",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 200
        }
      },
      "q": {
        "name": "q",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 120
        }
      },
      "from": {
        "name": "from",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "format": "date-time"
        }
      },
      "to": {
        "name": "to",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "format": "date-time"
        }
      },
      "stop_id": {
        "name": "stop_id",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 200
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "StopPage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Stop"
            }
          },
          "page": {
            "$ref": "#/components/schemas/Page"
          }
        },
        "required": [
          "data",
          "page"
        ]
      },
      "VehiclePage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vehicle"
            }
          },
          "page": {
            "$ref": "#/components/schemas/Page"
          }
        },
        "required": [
          "data",
          "page"
        ]
      },
      "TripPage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Trip"
            }
          },
          "page": {
            "$ref": "#/components/schemas/Page"
          }
        },
        "required": [
          "data",
          "page"
        ]
      },
...
      "HistoryPointPage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HistoryPoint"
            }
          },
          "page": {
            "$ref": "#/components/schemas/Page"
          }
        },
        "required": [
          "data",
          "page"
        ]
      },
...
    }
  }
}